+++ src/examples/server_utilisation.py
@@ -28,7 +28,7 @@ import time
 
 os.environ.setdefault("SQLALCHEMY_SILENCE_UBER_WARNING", "1")
 
-import questdb_connect.dialect as qdbc
+import questdb_connect as qdbc
 from sqlalchemy import Column, MetaData, create_engine
 from sqlalchemy.orm import Session, declarative_base
