+++ src/questdb_connect/_utils.py
@@ -1,0 +1,48 @@
+#
+#     ___                  _   ____  ____
+#    / _ \ _   _  ___  ___| |_|  _ \| __ )
+#   | | | | | | |/ _ \/ __| __| | | |  _ \
+#   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+#    \__\_\\__,_|\___||___/\__|____/|____/
+#
+#  Copyright (c) 2014-2019 Appsicle
+#  Copyright (c) 2019-2023 QuestDB
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+import re
+
+public_schema_filter = re.compile(
+    r"(')?(public(?(1)\1|)\.)", re.IGNORECASE | re.MULTILINE
+)
+
+
+def remove_public_schema(query):
+    if query and isinstance(query, str) and "public" in query:
+        return re.sub(public_schema_filter, "", query)
+    return query
+
+
+def quote_identifier(identifier: str):
+    if not identifier:
+        return None
+    first = 0
+    last = len(identifier)
+    if identifier[first] in _QUOTES:
+        first += 1
+    if identifier[last - 1] in _QUOTES:
+        last -= 1
+    return f'"{identifier[first:last]}"'
+
+
+_QUOTES = ("'", '"')
