+++ src/questdb_connect/compilers.py
@@ -1,0 +1,60 @@
+#
+#     ___                  _   ____  ____
+#    / _ \ _   _  ___  ___| |_|  _ \| __ )
+#   | | | | | | |/ _ \/ __| __| | | |  _ \
+#   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+#    \__\_\\__,_|\___||___/\__|____/|____/
+#
+#  Copyright (c) 2014-2019 Appsicle
+#  Copyright (c) 2019-2023 QuestDB
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+import abc
+
+import sqlalchemy
+
+from ._utils import quote_identifier, remove_public_schema
+from .types import QDBTypeMixin
+
+
+class QDBDDLCompiler(sqlalchemy.sql.compiler.DDLCompiler, abc.ABC):
+    def visit_create_schema(self, create, **kw):
+        raise Exception("QuestDB does not support SCHEMAS, there is only 'public'")
+
+    def visit_drop_schema(self, drop, **kw):
+        raise Exception("QuestDB does not support SCHEMAS, there is only 'public'")
+
+    def visit_create_table(self, create, **kw):
+        table = create.element
+        create_table = f"CREATE TABLE {quote_identifier(table.fullname)} ("
+        create_table += ", ".join(
+            [self.get_column_specification(c.element) for c in create.columns]
+        )
+        return create_table + ") " + table.engine.get_table_suffix()
+
+    def get_column_specification(self, column: sqlalchemy.Column, **_):
+        if not isinstance(column.type, QDBTypeMixin):
+            raise sqlalchemy.exc.ArgumentError(
+                "Column type is not a valid QuestDB type"
+            )
+        return column.type.column_spec(column.name)
+
+
+class QDBSQLCompiler(sqlalchemy.sql.compiler.SQLCompiler, abc.ABC):
+    def _is_safe_for_fast_insert_values_helper(self):
+        return True
+
+    def visit_textclause(self, textclause, add_to_result_map=None, **kw):
+        textclause.text = remove_public_schema(textclause.text)
+        return super().visit_textclause(textclause, add_to_result_map, **kw)
