+++ src/questdb_connect/types.py
@@ -22,8 +22,37 @@
 #
 import enum
 
-import sqlalchemy as sqla
-from sqlalchemy.exc import ArgumentError
+import sqlalchemy
+
+from ._utils import quote_identifier
+
+__all__ = [
+    "PartitionBy",
+    "QDBTypeMixin",
+    "Boolean",
+    "Byte",
+    "Short",
+    "Char",
+    "Int",
+    "Long",
+    "Date",
+    "Timestamp",
+    "Float",
+    "Double",
+    "String",
+    "Symbol",
+    "Long256",
+    "GeohashByte",
+    "GeohashShort",
+    "GeohashInt",
+    "GeohashLong",
+    "UUID",
+    "QUESTDB_TYPES",
+    "resolve_type_from_name",
+    "geohash_class",
+    "geohash_type_name",
+]
+
 
 # ===== QUESTDB PARTITION TYPE =====
 
@@ -39,7 +68,6 @@ class PartitionBy(enum.Enum):
 
 # ===== QUESTDB DATA TYPES =====
 
-_QUOTES = ("'", '"')
 _GEOHASH_BYTE_MAX = 8
 _GEOHASH_SHORT_MAX = 16
 _GEOHASH_INT_MAX = 32
@@ -52,7 +80,9 @@ _TYPE_CACHE = {
 
 def geohash_type_name(bits):
     if not isinstance(bits, int) or bits < 0 or bits > _GEOHASH_LONG_BITS:
-        raise ArgumentError(f"geohash precision must be int [0, {_GEOHASH_LONG_BITS}]")
+        raise sqlalchemy.exc.ArgumentError(
+            f"geohash precision must be int [0, {_GEOHASH_LONG_BITS}]"
+        )
     if 0 < bits <= _GEOHASH_BYTE_MAX:
         return f"GEOHASH(8b)"
     elif _GEOHASH_BYTE_MAX < bits <= _GEOHASH_SHORT_MAX:
@@ -64,7 +94,9 @@ def geohash_type_name(bits):
 
 def geohash_class(bits):
     if not isinstance(bits, int) or bits < 0 or bits > _GEOHASH_LONG_BITS:
-        raise ArgumentError(f"geohash precision must be int [0, {_GEOHASH_LONG_BITS}]")
+        raise sqlalchemy.exc.ArgumentError(
+            f"geohash precision must be int [0, {_GEOHASH_LONG_BITS}]"
+        )
     if 0 < bits <= _GEOHASH_BYTE_MAX:
         return GeohashByte
     elif _GEOHASH_BYTE_MAX < bits <= _GEOHASH_SHORT_MAX:
@@ -74,21 +106,9 @@ def geohash_class(bits):
     return GeohashLong
 
 
-def quote_identifier(identifier: str):
-    if not identifier:
-        return None
-    first = 0
-    last = len(identifier)
-    if identifier[first] in _QUOTES:
-        first += 1
-    if identifier[last - 1] in _QUOTES:
-        last -= 1
-    return f'"{identifier[first:last]}"'
-
-
-class QDBTypeMixin(sqla.types.TypeDecorator):
+class QDBTypeMixin(sqlalchemy.types.TypeDecorator):
     __visit_name__ = "QDBTypeMixin"
-    impl = sqla.types.String
+    impl = sqlalchemy.types.String
     cache_ok = True
 
     @classmethod
@@ -104,20 +124,20 @@ class QDBTypeMixin(sqla.types.TypeDecorator):
 
 class Boolean(QDBTypeMixin):
     __visit_name__ = "BOOLEAN"
-    impl = sqla.types.Boolean
+    impl = sqlalchemy.types.Boolean
     type_code = 1
 
 
 class Byte(QDBTypeMixin):
     __visit_name__ = "BYTE"
-    impl = sqla.types.Integer
+    impl = sqlalchemy.types.Integer
     type_code = 2
 
 
 class Short(QDBTypeMixin):
     __visit_name__ = "SHORT"
     type_code = 3
-    impl = sqla.types.Integer
+    impl = sqlalchemy.types.Integer
 
 
 class Char(QDBTypeMixin):
@@ -128,37 +148,37 @@ class Char(QDBTypeMixin):
 class Int(QDBTypeMixin):
     __visit_name__ = "INT"
     type_code = 5
-    impl = sqla.types.Integer
+    impl = sqlalchemy.types.Integer
 
 
 class Long(QDBTypeMixin):
     __visit_name__ = "LONG"
     type_code = 6
-    impl = sqla.types.Integer
+    impl = sqlalchemy.types.Integer
 
 
 class Date(QDBTypeMixin):
     __visit_name__ = "DATE"
     type_code = 7
-    impl = sqla.types.Date
+    impl = sqlalchemy.types.Date
 
 
 class Timestamp(QDBTypeMixin):
     __visit_name__ = "TIMESTAMP"
     type_code = 8
-    impl = sqla.types.DateTime
+    impl = sqlalchemy.types.DateTime
 
 
 class Float(QDBTypeMixin):
     __visit_name__ = "FLOAT"
     type_code = 9
-    impl = sqla.types.Float
+    impl = sqlalchemy.types.Float
 
 
 class Double(QDBTypeMixin):
     __visit_name__ = "DOUBLE"
     type_code = 10
-    impl = sqla.types.Float
+    impl = sqlalchemy.types.Float
 
 
 class String(QDBTypeMixin):
