--- superset_toolkit/docker/requirements-local.txt
@@ -1,1 +1,0 @@
-questdb-connect==0.0.41
